+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 
 ## (UNRELEASED)
 
+### FEATURES:
+* provisioner/WindowsShell: add support for packer's windows shell provisioner
+* provisioner/WindowsRestart: add support for packer's windows restart provisioner
 
 ## 0.2.2 (August 13, 2017)